Refactor Game Persistence and Server Code
- Updated GameDbContext to use EquipmentSlotEntity for better clarity.
- Changed DisconnectPersistenceDto and PositionState to use int for Direction instead of byte.
- Enhanced AccountCharacterService to handle character creation errors more gracefully and removed redundant calculations for max HP and MP.
- Adjusted AccountLoginService and AccountRegistrationService to include necessary entity references.
- Improved GameServer to manage character deletion and movement snapshots more effectively.
- Refined PlayerSpawnService to correctly initialize player stats and positions.
- Updated various classes in the GameECS project to ensure proper entity management and navigation systems.
- Introduced new MovementPackets and MovementInput components for better movement handling in the ECS.
- Created ServerGameSimulation to manage the game world and player interactions more efficiently.
- Adjusted client-side scripts to align with the new server structure and data handling.

+++ Game.Persistence/Persistence/DisconnectState.cs
@@ -11,7 +11,7 @@ public sealed record DisconnectPersistenceDto
     public required int PositionX { get; init; }
     public required int PositionY { get; init; }
     public required int PositionZ { get; init; }
-    public required byte Dir { get; init; }
+    public required int Dir { get; init; }
     public int CurrentHp { get; init; }
     public int CurrentMp { get; init; }
